+++ sdk/go/aiven/database.go
@@ -24,8 +24,8 @@ import (
 // 	pulumi.Run(func(ctx *pulumi.Context) error {
 // 		_, err := aiven.NewDatabase(ctx, "mydatabase", &aiven.DatabaseArgs{
 // 			DatabaseName: pulumi.String("<DATABASE_NAME>"),
-// 			Project:      pulumi.String(aiven_project.Myproject.Project),
-// 			ServiceName:  pulumi.String(aiven_service.Myservice.Service_name),
+// 			Project:      pulumi.Any(aiven_project.Myproject.Project),
+// 			ServiceName:  pulumi.Any(aiven_service.Myservice.Service_name),
 // 		})
 // 		if err != nil {
 // 			return err