+++ sdk/python/pulumi_aiven/get_kafka_acl.py
@@ -6,7 +6,8 @@ import warnings
 import pulumi
 import pulumi.runtime
 from typing import Union
-from . import utilities, tables
+from . import _utilities, _tables
+
 
 class GetKafkaAclResult:
     """
@@ -34,6 +35,8 @@ class GetKafkaAclResult:
         if username and not isinstance(username, str):
             raise TypeError("Expected argument 'username' to be a str")
         __self__.username = username
+
+
 class AwaitableGetKafkaAclResult(GetKafkaAclResult):
     # pylint: disable=using-constant-test
     def __await__(self):
@@ -47,13 +50,12 @@ class AwaitableGetKafkaAclResult(GetKafkaAclResult):
             topic=self.topic,
             username=self.username)
 
-def get_kafka_acl(permission=None,project=None,service_name=None,topic=None,username=None,opts=None):
+
+def get_kafka_acl(permission=None, project=None, service_name=None, topic=None, username=None, opts=None):
     """
     Use this data source to access information about an existing resource.
     """
     __args__ = dict()
-
-
     __args__['permission'] = permission
     __args__['project'] = project
     __args__['serviceName'] = service_name
@@ -62,7 +64,7 @@ def get_kafka_acl(permission=None,project=None,service_name=None,topic=None,user
     if opts is None:
         opts = pulumi.InvokeOptions()
     if opts.version is None:
-        opts.version = utilities.get_version()
+        opts.version = _utilities.get_version()
     __ret__ = pulumi.runtime.invoke('aiven:index/getKafkaAcl:getKafkaAcl', __args__, opts=opts).value
 
     return AwaitableGetKafkaAclResult(