+++ sdk/python/pulumi_aiven/service.py
@@ -6,7 +6,7 @@ import warnings
 import pulumi
 import pulumi.runtime
 from typing import Union
-from . import utilities, tables
+from . import _utilities, _tables
 
 
 class Service(pulumi.CustomResource):
@@ -74,6 +74,7 @@ class Service(pulumi.CustomResource):
         * `indicesQueriesCacheSize` (`str`)
         * `indicesQueryBoolMaxClauseCount` (`str`)
         * `reindexRemoteWhitelists` (`list`)
+        * `searchMaxBuckets` (`str`)
         * `threadPoolAnalyzeQueueSize` (`str`)
         * `threadPoolAnalyzeSize` (`str`)
         * `threadPoolForceMergeSize` (`str`)
@@ -285,6 +286,7 @@ class Service(pulumi.CustomResource):
         * `logRetentionHours` (`str`)
         * `logSegmentBytes` (`str`)
         * `maxConnectionsPerIp` (`str`)
+        * `maxIncrementalFetchSessionCacheSlots` (`str`)
         * `messageMaxBytes` (`str`)
         * `numPartitions` (`str`)
         * `offsetsRetentionMinutes` (`str`)
@@ -642,6 +644,7 @@ class Service(pulumi.CustomResource):
             * `indicesQueriesCacheSize` (`pulumi.Input[str]`)
             * `indicesQueryBoolMaxClauseCount` (`pulumi.Input[str]`)
             * `reindexRemoteWhitelists` (`pulumi.Input[list]`)
+            * `searchMaxBuckets` (`pulumi.Input[str]`)
             * `threadPoolAnalyzeQueueSize` (`pulumi.Input[str]`)
             * `threadPoolAnalyzeSize` (`pulumi.Input[str]`)
             * `threadPoolForceMergeSize` (`pulumi.Input[str]`)
@@ -827,6 +830,7 @@ class Service(pulumi.CustomResource):
             * `logRetentionHours` (`pulumi.Input[str]`)
             * `logSegmentBytes` (`pulumi.Input[str]`)
             * `maxConnectionsPerIp` (`pulumi.Input[str]`)
+            * `maxIncrementalFetchSessionCacheSlots` (`pulumi.Input[str]`)
             * `messageMaxBytes` (`pulumi.Input[str]`)
             * `numPartitions` (`pulumi.Input[str]`)
             * `offsetsRetentionMinutes` (`pulumi.Input[str]`)
@@ -1029,7 +1033,7 @@ class Service(pulumi.CustomResource):
         if not isinstance(opts, pulumi.ResourceOptions):
             raise TypeError('Expected resource options to be a ResourceOptions instance')
         if opts.version is None:
-            opts.version = utilities.get_version()
+            opts.version = _utilities.get_version()
         if opts.id is None:
             if __props__ is not None:
                 raise TypeError('__props__ is only valid when passed in combination with a valid opts.id to get an existing resource')
@@ -1182,6 +1186,7 @@ class Service(pulumi.CustomResource):
             * `indicesQueriesCacheSize` (`pulumi.Input[str]`)
             * `indicesQueryBoolMaxClauseCount` (`pulumi.Input[str]`)
             * `reindexRemoteWhitelists` (`pulumi.Input[list]`)
+            * `searchMaxBuckets` (`pulumi.Input[str]`)
             * `threadPoolAnalyzeQueueSize` (`pulumi.Input[str]`)
             * `threadPoolAnalyzeSize` (`pulumi.Input[str]`)
             * `threadPoolForceMergeSize` (`pulumi.Input[str]`)
@@ -1367,6 +1372,7 @@ class Service(pulumi.CustomResource):
             * `logRetentionHours` (`pulumi.Input[str]`)
             * `logSegmentBytes` (`pulumi.Input[str]`)
             * `maxConnectionsPerIp` (`pulumi.Input[str]`)
+            * `maxIncrementalFetchSessionCacheSlots` (`pulumi.Input[str]`)
             * `messageMaxBytes` (`pulumi.Input[str]`)
             * `numPartitions` (`pulumi.Input[str]`)
             * `offsetsRetentionMinutes` (`pulumi.Input[str]`)
@@ -1603,7 +1609,7 @@ class Service(pulumi.CustomResource):
         return Service(resource_name, opts=opts, __props__=__props__)
 
     def translate_output_property(self, prop):
-        return tables._CAMEL_TO_SNAKE_CASE_TABLE.get(prop) or prop
+        return _tables.CAMEL_TO_SNAKE_CASE_TABLE.get(prop) or prop
 
     def translate_input_property(self, prop):
-        return tables._SNAKE_TO_CAMEL_CASE_TABLE.get(prop) or prop
+        return _tables.SNAKE_TO_CAMEL_CASE_TABLE.get(prop) or prop